Create SeedMaker Version 0.0.3
GROG is on the job again. Better user instructions updated and Help menu and descriptions in the app and about. Some changes to menus, buttons etc and functionality in general. Created an SeedMaker.exe and packaged it all up and got GROG to write up instructions on how user can do the same.